requirements.txt
requests==2.33.0
pytz==2026.1
PyYAML==6.0.3
fastmcp==2.12.5
feedparser==6.0.12
boto3==1.42.76
litellm==1.82.6
json-repair==0.58.6
